requirements.txt
# ConsistentTutor - AI Tutor with Adaptive Instruction Intelligence (AII)

# Core LLM & RAG Framework
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.0.53
requests>=2.31.0

# Embeddings & Vector Database
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0

# PDF Processing & Image Extraction
pymupdf>=1.23.8
Pillow>=10.0.0

# Text Processing
PyYAML>=6.0

# Web Framework
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
